refactoring after class: using find and findIndex methods
