+++ account/src/use-case/changeUserName.js
@@ -3,10 +3,10 @@ import {searchUserAccountByEmailUseCase} from "./searchUserAccountByEmail.js";
 
 const changeUserNameUseCase = (email, newName) => {
     const userByEmail = searchUserAccountByEmailUseCase(email);
-    if (userByEmail == false){return false}
-    else if (userByEmail[0].name == newName){return false}
+    if (!userByEmail){return false}
+    else if (userByEmail.name == newName){return false}
     else{
-        const userIndex = userByEmail[0].id - 1;
+        const userIndex = usersArray.findIndex(user => user.email === email);
         usersArray[userIndex].name = newName;
         return true
     }
