+++ account/src/use-case/createUserAccount.js
@@ -1,16 +1,15 @@
-import {usersArray} from "../mockUsers.js"
+import {usersArray} from "../mockUsers.js";
 
-const createUserUseCase = (name, email, password) => {
+export function createUserUseCase (name, email, password){
     var newUser = {
         userId: usersArray.length + 1,
         name,
         email,
         password,
-        createdDate: new Date().toISOString().split('T')[0]
+        createdDate: new Date().toISOString().substring(0, 10)
     }
     usersArray.push(newUser)
     
     return newUser;
 };
 
-export {createUserUseCase};
