+++ account/src/use-case/removeUserAccount.js
@@ -2,14 +2,12 @@ import {usersArray} from "../mockUsers.js";
 import {searchUserAccountByEmailUseCase} from "./searchUserAccountByEmail.js";
 
 const removeUserUseCase = (email) => {
-    const userByEmail = searchUserAccountByEmailUseCase(email);  
-    if (userByEmail === false){
-        return false
-    }
+    const userByEmail = searchUserAccountByEmailUseCase(email);
+    if (!userByEmail){return false}
     else {
-        const userByEmailIndex = userByEmail[0].id - 1;
-        usersArray.splice(userByEmailIndex, 1);
-        return true 
+        const userIndex = usersArray.findIndex(user => user.email === email);
+        usersArray.splice(userIndex, 1);
+        return true
     }
     
 };
