+++ account/src/use-case/searchUserAccountByEmail.js
@@ -1,12 +1,7 @@
 import {usersArray} from "../mockUsers.js";
 
 const searchUserAccountByEmailUseCase = (email) => {
-    const userByEmail = usersArray.filter((user) => user.email === email)
-   
-    if (userByEmail.length > 0) {
-        return userByEmail}
-    else {
-        return false}
+    return usersArray.find((user) => user.email === email)
 };
 
 export {searchUserAccountByEmailUseCase};
