+++ account/test/filterUsersByUF.test.js
@@ -1,4 +1,3 @@
-import {usersArray} from "../src/mockUsers.js"; 
 import {filterUsersByUFUseCase} from "../src/use-case/filterUsersByUF.js";
 
 console.log(filterUsersByUFUseCase('SP'));
