+++ account/test/removeUserAccount.test.js
@@ -1,4 +1,3 @@
-import {usersArray} from "../src/mockUsers.js"; 
 import {removeUserUseCase} from "../src/use-case/removeUserAccount.js";
 
 console.log(removeUserUseCase('maria@email.com'));
