+++ account/test/searchUserAccountByEmail.test.js
@@ -1,4 +1,3 @@
-import {usersArray} from "../src/mockUsers.js"; 
 import {searchUserAccountByEmailUseCase} from "../src/use-case/searchUserAccountByEmail.js";
 
 console.log(searchUserAccountByEmailUseCase('joselima@email.com'));
